+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to
 #### Breaking Changes
 - Introduce scripts/dialog/main.sh;
 #### Added
+- glibc: test fread.c;
 - oom: introduce adaptive-oom-score tools;
 - gdb/samples: add mmap-file.gdb to test mmap syscall, add signal.gdb, fcntl.gdb, mman.gdb;
 #### Changed